test(e2e): split 703-LOC helpers.ts into peers/ui/touch modules
Per spec §'Helpers redesign' and PR-2 plan Tasks 7-10. Three focused
modules behind a re-export barrel — legacy specs import-from-helpers
unchanged, new specs import directly from the focused module they need.

Behaviour preserved verbatim. Magic-number sleeps and 30s timeout
overrides stay where they were; PR-2 only deletes them from the pilot
spec (multi-peer-sync.spec.ts) where the Peer wrapper replaces them.
Other 7 specs migrate file-by-file via tracking issue #458.

The barrel drops its eslint-disable header — re-exports contain no
waitForTimeout calls, and a lingering unused-disable warning would
trip CI.
